Added meta data for `cargo deb`
Now results `cargo deb` in a .deb being build.
The debian package has the binaries stalwart and stalwart-cli.

For what it is worth:
cargo-deb finds out that "stalwart" is the default rust package.
That rust package "stalwart-cli" is needed to be build is also
something that cargo-deb figures out.
To be able to do `cargo deb` you first have to do `cargo install cargo-deb`